category edit/delete error fix

+++ app/Http/Controllers/Backend/IndCategoryController.php
@@ -3,19 +3,21 @@
 namespace App\Http\Controllers\Backend;
 
 use App\Models\Category;
-use App\Models\ServiceType;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCategory;
 use App\Http\Requests\UpdateCategory;
+use Illuminate\Support\Facades\Storage;
 
-// TODO:: Some of these requests classes are empty, so fill these with whatever you can.
+// TODO:: Need to do request validation.
 
 class IndCategoryController extends Controller
 {
-
     public function index()
     {
-        $categories = Category::getAll('ind')->orderBy('updated_at', 'DSC')->paginate(15);
+        $categories = Category::getAll('ind')
+            ->orderBy('updated_at', 'DSC')
+            ->paginate(15);
+
         $navs = $this->navs();
 
         return view('backend.ind-category.index', compact('categories', 'navs'));
@@ -23,45 +25,59 @@ class IndCategoryController extends Controller
 
     public function store(StoreCategory $request)
     {
-        $category = new Category([
-            'name' => $request->post('category'),
-            'is_confirmed' => 1,
-            'image' => $request->file('image')->store('category-images')
-        ]);
-        ServiceType::where('name', 'ind')->first()->categories()->save($category);
+        $category = new Category();
+        $category->name = $request->input('category');
+        $category->service_type_id = 1;
+        $category->image = $request->file('image')->store('category-images');
+        $category->save();
 
-        return redirect(route('individual-category.show', $category->id))->with('success', 'Category "' . $request->post('category') . '" Added Successfully!');
+        return back()->with('success', $request->input('category').' ক্যাটাগরিটি তৈরি হয়েছে।');
     }
-
 
     public function show(Category $category)
     {
+        $subCategories = $category->subCategories('confirmed')
+            ->orderBy('updated_at', 'DSC')
+            ->paginate(10);
+
         $navs = $this->navs();
-        $subCategories = $category->subCategories('confirmed')->orderBy('updated_at', 'DSC')->paginate(10);
+
         return view('backend.ind-category.show', compact('category', 'subCategories', 'navs'));
     }
 
-
     public function update(UpdateCategory $request, Category $category)
     {
-        $category->update([
-            'name' => $request->post('category')
-        ]);
-        return back()->with('success', 'Category Renamed Successfully!');
+        $oldName = $category->name;
+
+        $category->name = $request->input('name');
+        if ($request->hasFile('image')) {
+            Storage::delete($category->image);
+            $category->image = $request->file('image')->store('category-images');
+        }
+        $category->save();
+
+        return back()->with('success', $oldName.' ক্যাটাগরিটি এডিট হয়েছে।');
     }
 
     public function destroy(Category $category)
     {
-        // TODO:: There can be some kind of gotcha about 'Cascade' deleting or updating, I'm not sure!.
-        // TODO:: Delete image
-        $category->delete();
-        return redirect(route('individual-category.index'))->with('success', 'Category Deleted Successfully!');
+        if ($category->subCategories->isNotEmpty()) {
+            return back()->with('error', $category->name.' ক্যাটাগরিটির ভিতর সাব-ক্যাটাগরি রয়েছে তাই এটি ডিলিট হবে না।');
+        }
+        else
+        {
+            $category->delete();
+            Storage::delete($category->image);
+
+            return back()->with('success', $category->name.' ক্যাটাগরিটি ডিলিট হয়েছে।');
+        }
     }
 
     private function navs()
     {
         return [
-            ['url' => route('individual-category.index'), 'text' => 'ব্যক্তি সেবা প্রদানকারী বিভাগসমূহ']
+            ['url' => route('individual-category.index'), 'text' => 'বেক্তিগত ক্যাটাগরি সমূহ'],
+            ['url' => route('organization-category.index'), 'text' => 'প্রাতিষ্ঠানিক ক্যাটাগরি সমূহ']
         ];
     }
 }
